[chore] 폴더 분리
